// frontend/vue-project/src/api/models/pageResultVirtualFileVO.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { VirtualFileVO } from './virtualFileVO';

/**
 * 响应数据
 */
export interface PageResultVirtualFileVO {
  total?: number;
  pages?: number;
  current?: number;
  size?: number;
  list?: VirtualFileVO[];
}